extract common function to util file, move tests
